2025-09-05 Long version of the RSL Pin Holder
Added some notches to the bottom of the pin holder edge to act as a place to anchor Tacking. The sample I provided to Maddy became dislodged when torqued and she suggested adding a way to tack the bottom of the pin as a way to prevent the issue. In addition the text identifying the piece is now embossed to prevent wear on the Type III webbing as it is used.